+++ test/unit_test/project_management_tool/service.py
@@ -5,7 +5,7 @@ from typing import Dict
 
 import urllib3
 
-from create_pr_bot.project_management_tool.clickup import ClickUpClient
+from create_pr_bot.project_management_tool.service import ClickUpClient
 
 
 @pytest.fixture
